+++ src/containers/Parsing/FilesAssembler.js
