+++ src/containers/Parsing/Loading
@@ -1,0 +1,50 @@
+export default (file, callbackProgressF, callbackEndF) => {
+  const CHUNK_SIZE = 10 * 1024;
+  let start = 0;
+  let end;
+  let slice;
+  const extension = file.name.slice(file.name.lastIndexOf('.') - 1 + 2);
+  if (file === undefined) {
+    return;
+  }
+  if (extension !== 'json' && extension !== 'geojson') {
+    alert('Wrong extension. Should be .json or .geojson');
+    return;
+  }
+  if (file.size === 0) {
+    return;
+  }
+  let reader = new FileReader();
+  reader.onload = function (evt) {
+    reader.offset = start;
+    callbackRead(this, file, evt, callbackProgressF, callbackEndF);
+    start += CHUNK_SIZE;
+    Load();
+  };
+  reader.size = CHUNK_SIZE;
+
+  Load();
+
+  function Load() {
+    if (start < file.size) {
+      end = min(start + CHUNK_SIZE, file.size);
+      slice = file.slice(start, end);
+      reader.readAsArrayBuffer(slice);
+      end = null;
+      slice = null;
+    }
+  }
+  function callbackRead(reader, file, evt, callbackProgressF, callbackEndF) {
+    if (reader.offset + reader.size < file.size) {
+      callbackProgressF(evt.target.result);
+    } else {
+      callbackEndF(evt.target.result);
+    }
+  }
+  function min(a, b) {
+    return a < b ? a : b;
+  }
+}
+
+
+
